Fixed tests with arm refactor
