+++ unity/Assets/Scripts/ArmController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 76395f7b622f7416baad079589607e9c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
