+++ unity/Assets/Scripts/KinovaArmAgentController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 55a12199225ab4115a0db2691e7b66c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
