--- unity/debug/kiana_intersect_2.json
@@ -1,3 +1,0 @@
-[
-    {"action": "RotateWristRelative", "yaw": 10, "disableRendering": false, "returnToStart": true, "speed": 1}
-]
